requirements.txt
asgiref==3.11.0
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
distlib==0.4.0
django==5.2.9
djangorestframework-simplejwt==5.5.1
djangorestframework==3.16.1
drf-nested-routers==0.95.0
filelock==3.20.0
idna==3.11
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
packaging==25.0
pip==21.1.2
pipenv==2025.0.4
platformdirs==4.5.0
PyJWT==2.10.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.4
tomli==2.0.1
typeguard==4.3.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wheel==0.36.2
zipp==3.19.2
